refactor: lift message list data to `ChatView`
To make the multiselect state available to the chat header,
where we're going to move the multiselect counter.
